serialize repeat count

+++ src/constants.rs
@@ -51,7 +51,7 @@ pub const ARRAY_START_TOKEN: char = '|';
 pub const ARRAY_END_TOKEN: char = '÷';
 pub const ARRAY_REPEAT_TOKEN: char = 'þ';
 pub const ARRAY_REPEAT_MANY_TOKEN: char = '^';
-pub const ARRAY_REPEAT_COUNT_THRESHOLD: u8 = 4;
+pub const ARRAY_REPEAT_COUNT_THRESHOLD: i64 = 4;
 pub const OBJECT_START_TOKEN: char = '{';
 pub const OBJECT_END_TOKEN: char = '}';
